requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
IMDbPY==6.6
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.2
MarkupSafe==1.1.1
python-dotenv==0.10.1
requests==2.21.0
selenium==3.141.0
soupsieve==1.8
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
